requirements.txt
aniso8601==9.0.1
arrow==0.17.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coloredlogs==15.0
constantly==15.1.0
cryptography==3.4.7
Django==3.2.6
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.30.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-FontAwesome==0.1.5
Flask-JWT-Extended==4.1.0
Flask-MQTT==1.1.1
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gitdb==4.0.5
GitPython==3.1.14
greenlet==1.0.0
humanfriendly==9.1
hyperlink==21.0.0
idna==2.10
image==1.5.33
incremental==21.3.0
install==1.3.4
invoke==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.20.1
openpyxl==3.0.7
paho-mqtt==1.5.1
pandas==1.2.3
passlib==1.7.4
Pillow==8.3.1
pycparser==2.20
PyJWT==2.0.1
pyserial==3.5
pyserial-asyncio==0.5
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
roslibpy==1.2.1
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
sqlparse==0.4.1
squaternion==0.3.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
xlrd==2.0.1
zope.interface==5.4.0
